Fix review findings, safer factory reset and CSRF protection (v1.10.0)
Bug fixes:
- Front button AP mode: start the background WiFi retry interval in
  startAP(), so the AP isn't torn down seconds after a long press
- Runtime: getTotalRuntimeMinutes() no longer double-counts minutes that
  were already saved to storage; /api/status uses the same value as MQTT
- Web routes: register /api/status/lite and /api/diagnostic/buttons before
  their prefixes (AsyncWebServer prefix matching sent them to the wrong
  handler); static files are served after all API routes
- Web OTA: failed/stalled uploads are aborted and reset otaInProgress;
  30s inactivity timeout; upload state is tied to the owning request
- Filesystem upload: unmount FS and pause log writes during upload
- LED: night brightness 0% is no longer overridden to 50%; daytime
  brightness is LED_DEFAULT_BRIGHTNESS (100%) with or without night mode
- Default OTA password is OTA_PASSWORD ("diffuser-ota") as documented
- WiFi: detect SDK auto-reconnect while in DISCONNECTED state
- MQTT: length validation matches storage buffers; backoff capped at 60s
- Correct filesystem image name on the "files missing" page

Improvements:
- Rear button: 1s hold shows a red slow-blink warning (release cancels),
  5s hold confirms with a fast red blink and performs the factory reset
- CSRF: reject state-changing requests whose Origin/Referer doesn't match
  the Host (async server and ESP8266 safe update server)
- ESP32: recursive mutex (state_lock) serializes HTTP handlers with the
  main loop and MQTT command handling
- ESP8266 safe update mode: turn fan off, restart after 10 min idle
- Debounced fan speed persistence (5s) instead of a flash write per step
- Calibration and web OTA go through turnOff()/notifyStateChange()
- Remove unused MQTT/web callbacks and config constants; fix printf formats

+++ src/update_checker.cpp
@@ -117,13 +117,13 @@ void UpdateChecker::performCheck() {
 #ifdef PLATFORM_ESP8266
     // ESP8266: Check if we have enough memory for BearSSL TLS handshake
     uint32_t freeHeap = ESP.getFreeHeap();
-    logger.infof("Free heap for update check: %lu bytes", freeHeap);
+    logger.infof("Free heap for update check: %lu bytes", (unsigned long)freeHeap);
 
     // BearSSL needs ~12-15KB, plus the 1.5KB JSON doc and HTTP buffers (~2KB),
     // so 18KB is a more realistic floor than 15KB.
     if (freeHeap < 18000) {
-        snprintf(_info.errorMessage, sizeof(_info.errorMessage), "Low memory (%lu bytes)", freeHeap);
-        logger.warnf("Update check skipped: only %lu bytes free", freeHeap);
+        snprintf(_info.errorMessage, sizeof(_info.errorMessage), "Low memory (%lu bytes)", (unsigned long)freeHeap);
+        logger.warnf("Update check skipped: only %lu bytes free", (unsigned long)freeHeap);
         _state = UpdateCheckState::ERROR;
         if (_stateCallback) _stateCallback();
         return;
